deep__search/pyproject.toml
[project]
name = "deep-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.9",
    "tavily-python>=0.7.11",
]
